Fixes & Updates.
No longer crash alongside a failed creeper command. I'm not even sure
why this was the default functionality. I must have been very tired. :\

More wine next time...

Updated the time flag to use .Duration for ease of use. @TODO is improve
the documentation around that so it's understood what is going on for that
and how to properly use it.

Changed the packaging of the arguments so it actually works this time. This
will probably need to be changed as I'm not happy with the switching of args
like that.